Removing some duplication in _grid-helpers.scss
It just refactors the sass to remove some duplication. The resulting css
should be unchanged (in fact after running grunt, the css indeed did not
change). This is to ensure that my planed changes to this file are
simpler.

Note - I'm not sure why the minified css ended up changing. Was it out
of date or am I missing something?